# ggml-cpu: Faster IQ1 mul_mat_vec on AVX2 using BMI2 instructions