@slaren 2025-03-02
> Is it fine to simply use the `__BMI2__` since "NATIVE" build is now the standard?

Please add also an option to enable it manually, add a check in [cpu-feats-x86.cpp](https://github.com/ggml-org/llama.cpp/blob/master/ggml/src/ggml-cpu/cpu-feats-x86.cpp), and add it to the CPU variant list in:

https://github.com/ggml-org/llama.cpp/blob/cc473cac7cea1484c1f870231073b0bf0352c6f9/ggml/src/CMakeLists.txt#L308-L312

You could also check for Zen 2 in `cpu-feats-x86.cpp`, and if necessary add a variant for Zen 2 that excludes this feature.