@remyoudompheng 2025-03-04
Update with CMakeLists changes (no Zen 2 specific case, maybe a separate PR can add AMD microarchitectures).