@slaren 2025-03-04
Looks good, thanks.

It would also be necessary to add a `ggml_cpu_has_bmi2` function and report it in `ggml_backend_cpu_get_features`:
https://github.com/ggml-org/llama.cpp/blob/1a24c4621f0280306b0d53a4fa474fc65d3f1b2e/ggml/src/ggml-cpu/ggml-cpu.cpp#L488

I suspect that MSVC will enable BMI2 with `/arch:avx2` or higher. After you make this change, then you can check in the "system info" string if BMI2 is being enabled. If so, then the definition would also need to be added here:
https://github.com/ggml-org/llama.cpp/blob/a3db57565fde170310ac10dc8bcef00f03bb6dfc/ggml/src/ggml-cpu/CMakeLists.txt#L209-L212
I can check for you if you don't have access to a machine with MSVC.