@remyoudompheng 2025-03-05
Done.
For MSVC it seems that it needs a manual define like AVXVNNI (tests on [godbolt.org](https://godbolt.org/z/fWbzTr9Ma) suggest that it always compiles intrinsics regardless of the `/arch` flag)