@slaren 2025-03-06
13900k:
| Model                       |   Threads | Test   |   t/s master |   t/s optim-x86 |   Speedup |
|:----------------------------|----------:|:-------|-------------:|----------------:|----------:|
| llama 8B IQ1_M - 1.75 bpw   |         8 | pp128  |        15.84 |           21.98 |      1.39 |
| llama 8B IQ1_M - 1.75 bpw   |         8 | tg32   |        13.45 |           17.31 |      1.29 |
| llama 8B IQ1_M - 1.75 bpw   |        16 | pp128  |        19.43 |           30.45 |      1.57 |
| llama 8B IQ1_M - 1.75 bpw   |        16 | tg32   |        16.81 |           22.04 |      1.31 |
| llama 8B IQ1_M - 1.75 bpw   |        24 | pp128  |        25.88 |           28.29 |      1.09 |
| llama 8B IQ1_M - 1.75 bpw   |        24 | tg32   |        18.73 |           23.52 |      1.26 |
| llama 8B IQ1_M - 1.75 bpw   |        32 | pp128  |        30.42 |           34.84 |      1.15 |
| llama 8B IQ1_M - 1.75 bpw   |        32 | tg32   |        19.85 |           24.61 |      1.24 |
| llama 8B IQ1_S - 1.5625 bpw |         8 | pp128  |        19.30 |           29.50 |      1.53 |
| llama 8B IQ1_S - 1.5625 bpw |         8 | tg32   |        17.57 |           21.46 |      1.22 |
| llama 8B IQ1_S - 1.5625 bpw |        16 | pp128  |        32.00 |           41.51 |      1.30 |
| llama 8B IQ1_S - 1.5625 bpw |        16 | tg32   |        22.62 |           27.48 |      1.21 |
| llama 8B IQ1_S - 1.5625 bpw |        24 | pp128  |        35.53 |           46.19 |      1.30 |
| llama 8B IQ1_S - 1.5625 bpw |        24 | tg32   |        24.82 |           27.44 |      1.11 |
| llama 8B IQ1_S - 1.5625 bpw |        32 | pp128  |        41.95 |           52.90 |      1.26 |
| llama 8B IQ1_S - 1.5625 bpw |        32 | tg32   |        25.55 |           26.71 |      1.05 |