@sandrohanea 2025-03-20
Hello @slaren , @remyoudompheng ,

It seems that after this PR x86 with AVX2 build for MSVC is failing:

![image](https://github.com/user-attachments/assets/6bc62c42-ba8d-4eed-acd7-c785132bff9b)
https://github.com/sandrohanea/whisper.net/actions/runs/13965684322/job/39095442481


cmake command:
```
cmake -S . -DGGML_NATIVE=OFF -A Win32 -DGGML_AVX=ON -DGGML_AVX2=ON -DGGML_FMA=ON -DGGML_F16C=ON -B build/win-x86
```

Do you have any recommendation on how to fix this issue?