@sandrohanea 2025-03-20
> Hello @slaren , @remyoudompheng ,
> 
> It seems that after this PR x86 with AVX2 build for MSVC is failing:
> 
> ![image](https://private-user-images.githubusercontent.com/40202887/424898928-6bc62c42-ba8d-4eed-acd7-c785132bff9b.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDI1MDIzMzUsIm5iZiI6MTc0MjUwMjAzNSwicGF0aCI6Ii80MDIwMjg4Ny80MjQ4OTg5MjgtNmJjNjJjNDItYmE4ZC00ZWVkLWFjZDctYzc4NTEzMmJmZjliLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTAzMjAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwMzIwVDIwMjAzNVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTlhZTg0ZWIxMjZkYTI1OTM5OGY0MzdlMmI4YzZiZjkwODM3MGRhOTc3ZjI1YWZhNDU2OWJkOTZjOTlhNzM1ZWUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.SnjvsOh97o0QFxb-3y9sLJe03QJPHf8a2-jdFuGZhx8) https://github.com/sandrohanea/whisper.net/actions/runs/13965684322/job/39095442481
> 
> cmake command:
> 
> ```
> cmake -S . -DGGML_NATIVE=OFF -A Win32 -DGGML_AVX=ON -DGGML_AVX2=ON -DGGML_FMA=ON -DGGML_F16C=ON -B build/win-x86
> ```
> 
> Do you have any recommendation on how to fix this issue?

Nevermind, just disabled the support for BMI2 on Win32 using `-DGGML_BMI2=OFF`.